+++ src/main/kotlin/club/sk1er/elementa/components/input/AbstractTextInput.kt
@@ -201,7 +201,7 @@ abstract class AbstractTextInput(
             val heightDifference = getHeight() - visualLines.size * 9f
             if (heightDifference > 0)
                 return@onMouseScroll
-            targetVerticalScrollingOffset = (targetVerticalScrollingOffset + it.delta * 9f).coerceIn(heightDifference, 0f)
+            targetVerticalScrollingOffset = (targetVerticalScrollingOffset + it.delta.toFloat() * 9f).coerceIn(heightDifference, 0f)
             it.stopPropagation()
         }
